put all code into llava dir
