+++ examples/llava/assets/xiaomi14pro_test.jpeg
