+++ examples/llava/minicpmv-convert-image-encoder-to-gguf.py
