+++ examples/llava/minicpmv_wrapper.cpp
@@ -1,7 +1,7 @@
 #include "ggml.h"
 #include "common.h"
 #include "clip.h"
-#include "minicpmv.h"
+#include "llava.h"
 #include "minicpmv_wrapper.h"
 #include "llama.h"
 #include <cstdio>
