+++ examples/llava/requirements.txt
@@ -1,3 +1,4 @@
 -r ../../requirements/requirements-convert.txt
 pillow~=10.2.0
 torch~=2.1.1
+torchvision==0.16.2
